Rename models.py to schema.py
Rename file and update import
